Added peliminary version of ticket contract
`ticket_contracts.xsd` contains the message contracts that will be used
to communicate with Qnet's Azure Servicebus Topics. It contains
references to types in `ticket_types.xsd`

`ticket_contracts.cs` is generated from  `ticket_contracts.xsd` and
`ticket_types.xsd` with the command `xsd /c ticket_contracts.xsd
ticket_types.xsd` and can be used to deserialize xml messages that are
based on ticket_contracts.xsd.

We'll document this more when we've advanced further in the dev cycle.