Added field ticket_guid. Updated note.type from enum to string
ticket_guid is used as an identifier for a ticket before the ticket is
created to be able to correlate the newly created ticket with an
external ticket. External parties will create a new guid and use this
value to match a newly created Marval ticket.

note.type is now a string instead of an enum, makes it easier to
understand when serialized. Values can be SOLUTION or HISTORY.